Refine placement of some symbols
